% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{add_to_corpus}
\alias{add_to_corpus}
\title{Add specified file to an existing corpus}
\usage{
add_to_corpus(corpus, f)
}
\arguments{
\item{corpus}{an existing hdf corpus to add to}

\item{f}{filename of hdf to be added to corpus}
}
\description{
Add specified file to an existing corpus
}

