% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{append_meta}
\alias{append_meta}
\title{Function appends to an existing ts1 or ts2 dataframe with the data queried from hec_metadata}
\usage{
append_meta(df)
}
\arguments{
\item{df}{a ts1 or ts2 dataframe}
}
\description{
Function appends to an existing ts1 or ts2 dataframe with the data queried from hec_metadata
}
\examples{
\dontrun{
# assuming x is a dataframe obtained from calling either extract_ts1 or extract_ts2
x \%>\% append_meta()
}
}
