% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{available_time_series}
\alias{available_time_series}
\title{List Available Time Series Data}
\usage{
available_time_series(f, area_name, domain = NULL, ...)
}
\arguments{
\item{f}{a hec_collection object read in with hec_file()}

\item{area_name}{the flow area name defined in the model}

\item{domain}{either 'one' or 'two' corresponding to one-d or two-d domains}
}
\description{
list available time series in the model that can be extracted with 
one of the extract_* functions
}
