% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{close_corpus}
\alias{close_corpus}
\title{Safely close a corpus}
\usage{
close_corpus(corp)
}
\arguments{
\item{corp}{corpus of hdf files read in with create_hdf_corpus}
}
\description{
Safely close a corpus
}

