% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{create_hdf_corpus}
\alias{create_hdf_corpus}
\title{Build a corpus of hdf files to execute queries on}
\usage{
create_hdf_corpus(path)
}
\arguments{
\item{path}{directory path containing hdf files to build corpus from}
}
\value{
list of files read in with hec_file
}
\description{
Build a corpus of hdf files to execute queries on
}
