% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-dim.R
\name{extract_ts1}
\alias{extract_ts1}
\title{Function queries an hdf file read in with hec_file for cross section time series data.}
\usage{
extract_ts1(f, station_name, ts_type = "Water Surface", timestamp = NULL)
}
\arguments{
\item{f}{an hdf5 file read in via hec_file()}

\item{station_name}{station(s) for the cross section to query time series from. 
Multiple cross sections must be passed as a vector.}

\item{ts_type}{time series to query out (ex 'Water Surface', 'Depth', ...)}

\item{timestatmp}{an optional timestamp to query, default action will not filter for a timestamp.
Use this option to speed up long queries when a desired timestamp is known.}
}
\description{
Function queries an hdf file read in with hec_file for cross section time series data.
}

