% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{extract_ts2}
\alias{extract_ts2}
\title{Function extracts a time series from an hdf file resulting 
from a HecRas model run. ts2 version is used for 2D portions of a Hec-Ras model.}
\usage{
extract_ts2(.f, x, y, ts_type = "Water Surface")
}
\arguments{
\item{.f}{an hdf file read in with hec_file or h5::h5file}

\item{x}{coordinate to query for}

\item{y}{coordinate to query for}

\item{ts_type}{the time series to extract}
}
\value{
dataframe for desired timeseries with relevant column attributes
}
\description{
Function extracts a time series from an hdf file resulting 
from a HecRas model run. ts2 version is used for 2D portions of a Hec-Ras model.
}
