% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{get_center_coordinates}
\alias{get_center_coordinates}
\title{Function retrieves the center coordinates for a given hec_file}
\usage{
get_center_coordinates(.f)
}
\arguments{
\item{.f}{a hdf file read in with \code{hec_file}}
}
\description{
Function retrieves the center coordinates for a given hec_file
}
