% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{get_flow_area_name}
\alias{get_flow_area_name}
\title{Function returns the path to 2D flow area defined in the hdf file.}
\usage{
get_flow_area_name(hf)
}
\arguments{
\item{hf}{an hdf5 object read in with either hec_file() or h5::h5file()}
}
\value{
a slash delmited path to the 2D flow area
}
\description{
Function returns the path to 2D flow area defined in the hdf file.
}

