% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{get_flow_area_path}
\alias{get_flow_area_path}
\title{Function returns the path to 2D flow area defined in the hdf file.}
\usage{
get_flow_area_path(hf)
}
\arguments{
\item{hf}{a hdf file read in with hec_file}
}
\value{
a slash delmited path to the 2D flow area
}
\description{
Function returns the path to 2D flow area defined in the hdf file.
}
