% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_timestamps}
\alias{get_model_timestamps}
\title{Get model timestamps 
This should correspond with the width of raw time series dataframes}
\usage{
get_model_timestamps(f)
}
\description{
Get model timestamps 
This should correspond with the width of raw time series dataframes
}
