% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{get_nearest_cell_center}
\alias{get_nearest_cell_center}
\title{Function returns the column index of the cell center coordinate nearest (x, y)}
\usage{
get_nearest_cell_center(x, y, nodes)
}
\arguments{
\item{x}{coordinate}

\item{y}{coordinate}

\item{nodes}{set of nodes to calculate distance from}
}
\description{
Function returns the column index of the cell center coordinate nearest (x, y)
}
