% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{get_nearest_cell_center_index}
\alias{get_nearest_cell_center_index}
\title{Function returns the column index of the cell center coordinate nearest to (x, y).
Uses simple vectorized version of Euclid's distance formula.}
\usage{
get_nearest_cell_center_index(x, y, nodes)
}
\arguments{
\item{x}{coordinate}

\item{y}{coordinate}

\item{nodes}{set of nodes to calculate distance from}
}
\description{
Function returns the column index of the cell center coordinate nearest to (x, y).
Uses simple vectorized version of Euclid's distance formula.
}

