% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{get_plan_attributes}
\alias{get_plan_attributes}
\title{Function retrieves model attributes embeded in a HEC-RAS result file.}
\usage{
get_plan_attributes(.f)
}
\arguments{
\item{.f}{an hdf5 file result from HEC-RAS}
}
\description{
Function retrieves model attributes embeded in a HEC-RAS result file.
}
