% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-dim.R
\name{get_xs_reach}
\alias{get_xs_reach}
\title{Function retrives the reach defined by the station name}
\usage{
get_xs_reach(.f, station_name)
}
\arguments{
\item{.f}{an hdf5 file read in with hdf_file or h5::h5file}

\item{station_name}{name of station to query reach name for}
}
\description{
Function retrives the reach defined by the station name
}
