% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-dim.R
\name{get_xs_river_stations}
\alias{get_xs_river_stations}
\title{Function returns cross section stations defined in the model}
\usage{
get_xs_river_stations(.f)
}
\arguments{
\item{.f}{an hdf5 file read in with hec_file or h5::h5file}
}
\description{
Function returns cross section stations defined in the model
}
