% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-dim.R
\name{get_xs_station_index}
\alias{get_xs_station_index}
\title{Function returns the index for a desired station name}
\usage{
get_xs_station_index(.f, station_name)
}
\arguments{
\item{.f}{an hdf5 file read in with hec_file or h5::h5file}

\item{station_name}{name of the station to query data for}
}
\description{
Function returns the index for a desired station name
}
