% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{hdf_paths}
\alias{hdf_paths}
\title{Constants point to location on the hdf5 file}
\format{An object of class \code{list} of length 6.}
\usage{
hdf_paths
}
\description{
Constants point to location on the hdf5 file
}
\keyword{datasets}
