% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{hec_attributes}
\alias{hec_attributes}
\title{Function retrieves top level metadata associated with a hec ras file}
\usage{
hec_attributes(hf)
}
\arguments{
\item{hf}{a \code{hec_file} object}
}
\value{
list of attributes associated with hecras output file
}
\description{
Function retrieves top level metadata associated with a hec ras file
}
