% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-two.R
\name{hec_coordinates}
\alias{hec_coordinates}
\title{Show center coordinates}
\usage{
hec_coordinates(hc)
}
\arguments{
\item{hc}{a hec_collection object}
}
\description{
show the center coordinates for items in a collection
}
