% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-one.R
\name{hec_crosssections}
\alias{hec_crosssections}
\title{Plan Cross Sections}
\usage{
hec_crosssections(hc)
}
\arguments{
\item{hc}{a hec_collection}
}
\value{
data frame
}
\description{
Extract the crossections for a given hec collection
}
