% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{hec_datasets}
\alias{hec_datasets}
\title{List Available data}
\usage{
hec_datasets(hc, domain = NULL)
}
\arguments{
\item{hc}{a "hec" object}

\item{domain}{which domain to list data for, will default to both, options are 'one' or 'two'}
}
\description{
list the datasets available in the hdf file. Use domain to show only one of 
'two' dimensions or 'one' crossetions. By default it will show both (if available).
}
