% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-one.R
\name{hec_extract.hec_one}
\alias{hec_extract.hec_one}
\title{Extract a time series from One D}
\usage{
\method{hec_extract}{hec_one}(hc, station_name, ts_type, time_stamp = NULL)
}
\arguments{
\item{hc}{a hec one collection}

\item{station_name}{one or more stations names}

\item{ts_type}{a time series type}

\item{time_stamp}{optional timestamp to filter query by}
}
\description{
describe this
}
