% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{hec_file2}
\alias{hec_file2}
\title{Function reads in either a single hdf file or a collection of these specified as arguments. 
It incorporates both hec_file and hdf_corpus into one function call from previous releases.}
\usage{
hec_file2(path, plan_numbers = NULL)
}
\arguments{
\item{path}{directory path to either a single hdf file or a directory of a collection of these}

\item{plan_numbers}{a vector of plan number associated with hdf files. For use when path is a directory.
Default action is to read all hdf files in path.}
}
\value{
list of files read in with hec_file
}
\description{
Function reads in either a single hdf file or a collection of these specified as arguments. 
It incorporates both hec_file and hdf_corpus into one function call from previous releases.
}
