% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{hec_info}
\alias{hec_info}
\title{Hec attributes}
\usage{
hec_info(hc)

hec_info(hc)
}
\arguments{
\item{hc}{a hec objet}

\item{hc}{a hec objet}
}
\value{
list of attributes

list of attributes
}
\description{
get all attributes relating to a hec object

get all attributes relating to a hec object
}
