% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-one.R
\name{hec_ls.hec_one}
\alias{hec_ls.hec_one}
\title{List Cross Section Data}
\usage{
\method{hec_ls}{hec_one}(hc)
}
\arguments{
\item{hc}{a hec file}
}
\value{
vector of data names
}
\description{
list all data available to query from either one or two dimensional domain
}
