% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecr.R
\name{hec_ls}
\alias{hec_ls}
\title{Print list of datasets}
\usage{
hec_ls(x, ...)
}
\description{
given a hec_collection object passed through one of 'hec_one' or 'hec_two' 
print the available data sets to query
}
