% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{hec_metadata}
\alias{hec_metadata}
\title{Function takes in a hec file or a collection of these and forms a metadata
dataframe from them.}
\usage{
hec_metadata(f)
}
\arguments{
\item{f}{a hec file object or collection of these}
}
\description{
Function takes in a hec file or a collection of these and forms a metadata
dataframe from them.
}
