% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-query.R
\name{hec_query}
\alias{hec_query}
\title{Query one dimensional domains}
\usage{
hec_query(hc, station_name, ts_type, time_stamp = NULL)
}
\arguments{
\item{hc}{a hec object. See ?hec_file for more.}

\item{station_name}{name(s) for station(s) defined in the model run. See ?hec_crosssections for more.}

\item{ts_type}{a valid time series type defined in the model run}
}
\description{
provide station(s), a type of time series and optionally a timestamp
to query out data from an hdf5 file resulting from a HEC-RAS model run
}
\examples{
\dontrun{
h <- hec_file("~/Docs/hec-model-run.hdf")
data <- hec_one(hc=h, station="19135.6", ts_type="Water Surface")
}
}
