% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{hec_timestamps}
\alias{hec_timestamps}
\title{Get timestamps}
\usage{
hec_timestamps(hc, time_zone = "UTC")
}
\arguments{
\item{hc}{a "hec" object}

\item{time_zone}{a character string (e.g., "UTC")}
}
\description{
Get timestamps
}
