% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-dim.R
\name{hec_ts1}
\alias{hec_ts1}
\title{Query one dimensional domains}
\usage{
hec_ts1(f, station_name, ts_type = "Water Surface", time_stamp = NULL)
}
\arguments{
\item{f}{a hec_collection object produced by calling hec_file()}

\item{station_name}{name(s) for station(s) defined in the model run}

\item{ts_type}{a valid time series type defined in the model run}
}
\description{
provide station(s), a type of time series and optionally a timestamp
to query out data from an hdf5 file resulting from a HEC-RAS model run
}
