% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-two.R
\name{hec_two}
\alias{hec_two}
\title{Query 2d Model Results}
\usage{
hec_two(hc, xy, ts_type = "Water Surface", time_stamp = NULL, time_zone = NULL)
}
\arguments{
\item{hc}{an hdf file read in with the hec_file() function}

\item{xy}{a coordinate or set of coordinates either in a dataframe or matrix structure. See
details below for more information.}

\item{ts_type}{the time series to extract}

\item{time_stamp}{return only values with this timestamp}

\item{time_zone}{character string (e.g.., "UTC")}
}
\description{
hec_two allows you query a time series from a hec-ras model result
using one or more coordinates and an optional time stamp.
}
\details{
You can supply coordinates in as a matrix. The matrix must have two columns, 
the first corresponding to the x the second to y. You can supply a dataframe 
with two columns, the first for x and the second for y.
}
\examples{
\dontrun{
## first read in file
f <- hec_file("examples/ArdenwoodCreek.p50.hdf")

## water surface time series at the coordinate 4567654.0, 2167453.0
ws <- hec_two(f, xy=c(4567654.0, 2167453.0), "Water Surface")

## water surface time series at multiple coordinates
coords <- c(4567654.0, 2167453.0, 3456124.0, 7856124.0)
ws <- hec_two(f, xy=coords, "Water Surface")

## water surface for a fixed timestamp, useful when querying for large amounts of coordinates. 
ws <- hec_two(f, xy=c(4567654.0, 2167453.0), "Water Surface", timestamp="2005-09-12 00:00:00")
}
}
