% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecr.R
\docType{package}
\name{hecr-package}
\alias{hecr-package}
\title{\code{hecr} package}
\description{
hecr is a simple package that allows a user to interface with an hdf5 file resulting from a HecRas model run. The api allows users to extract all important data from a model file without writting repetitive code. Data extraction follow the tidy data philosophy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FlowWest/hecr}
  \item Report bugs at \url{https://github.com/FlowWest/hecr/issues}
}

}
\author{
\strong{Maintainer}: emanuel rodriguez \email{erodriguez@flowwest.com}

}
