% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{is_hec_file}
\alias{is_hec_file}
\title{Check whether an object is of hec_file class and not a null pointer}
\usage{
is_hec_file(f)
}
\description{
Check whether an object is of hec_file class and not a null pointer
}

