% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-config.R
\name{new_config}
\alias{new_config}
\title{Create a new configuration file can be used to automate alot of the work 
done by the hecr package.}
\usage{
new_config(name, template = "default", in_wd = TRUE)
}
\arguments{
\item{name}{name to be given to the configuration}

\item{template}{template to use for the configuration, defaults to a 'default' template}
}
\description{
Create a new configuration file can be used to automate alot of the work 
done by the hecr package.
}

