% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{plan_info}
\alias{plan_info}
\title{Get Plan Information}
\usage{
plan_info(hc)
}
\arguments{
\item{hc}{a set of hec files read in with hec_file()}
}
\value{
tibble with attributes as columns and unique plans as rows
}
\description{
Given a plan or a collection of these return a set of metadata attributes
defined for each plan.
}
