% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_points_to_lines_geos.R
\name{snap_points_to_lines_geos}
\alias{snap_points_to_lines_geos}
\title{A function to snap sf POINT data to LINESTRING}
\usage{
snap_points_to_lines_geos(
  sf_points,
  sf_lin_net,
  snap_distance = 1000,
  vertex_density = 50,
  verbose = FALSE
)
}
\arguments{
\item{sf_points}{An sf POINT object.}

\item{sf_lin_net}{An sf LINESTRING object.}

\item{snap_distance}{numeric. Buffer distance around LINESTRING within which
to search for intersecting points.}

\item{vertex_density}{Density of vertices assigned to LINESTRING. Smaller
values lead to more precise snapping, which is increasingly useful with
increased line longitudinal resolution.}

\item{verbose}{Boolean. If TRUE, returns messages during processing.}
}
\value{
An sf POINT object of snapped points with joined features from lines
  and points. Points outwith buffer distance from the nearest line are
  discarded.
}
\description{
The function is used to snap POINT data to the nearest
  LINESTRING.
}
