% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{tree}
\alias{tree}
\title{Show tree directory of hdf file}
\usage{
tree(f, depth = 2)
}
\arguments{
\item{f}{a hec object}

\item{depth}{how deep to print the tree? Currently capped at 4}
}
\description{
Currently an experimental function used for visualizing 
the data structure in the hdf file.
}
