% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{create_hdf_corpus}
\alias{create_hdf_corpus}
\title{Build a corpus of hdf files to execute queries on. This function has been deprecated use hec_file2() instead}
\usage{
create_hdf_corpus(path, files = NULL)
}
\arguments{
\item{path}{directory path containing hdf files to build corpus from}

\item{files}{a vector of plan number associated with hdf files. When NULL (default) all hdf files in directory are read.}
}
\value{
list of files read in with hec_file
}
\description{
Build a corpus of hdf files to execute queries on. This function has been deprecated use hec_file2() instead
}

