% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thalweg_through_2D_domain.R
\name{extract_thalweg_through_2D_domain}
\alias{extract_thalweg_through_2D_domain}
\title{Extract thalweg line or}
\source{
\url{https://stat.ethz.ch/pipermail/r-sig-geo/2013-June/018449.html}
}
\usage{
extract_thalweg_through_2D_domain(
  hc,
  crs,
  number_of_sections,
  time_stamp = NULL,
  verbose = FALSE
)
}
\arguments{
\item{hc}{an hdf file read in with the hec_file() function}

\item{crs}{character string (e.g., "EPSG:3035)}

\item{number_of_sections}{number of cross sections to use to derive thalweg}

\item{time_stamp}{a timestamp length = 1 POSIXct object}

\item{verbose}{boolean}
}
\value{
A list of sfc POINT and LINESTRING objects. "thalweg_points",
  "thalweg_line", and "transect_lines"
}
\description{
HEC-RAS exports computational cell centres and cell elevations
  to hdf. We can use these to identify a thalweg through the 2D model domain.
  This function can return either the cell points as sfc POINTs, or sfc
  LINESTRING built from these points. The motivation for this function is to
  facilitate testing of HEC-RAS model outputs along the deepest stretch of
  the river. For example, we can check the attenuation of a flood wave from
  upstream to downstream through a 2D domain. This function uses Barry
  Rowlingson's code provied on the r-sig-geo mailing list on 5 June 2013 (see
  reference).
}
\details{
Note, only one thalweg will be generated, which can lead to
  unexpected results in cases of multi-channel (e.g.., braided) 2D model
  domains.
}
\examples{
path = system.file("extdata\\\\chippewa.hdf",package = "hecr")
h <- hecr::hec_file(path)

crs_prj <- sf::read_sf(system.file("extdata\\\\profile_lines.shp",
                                   package = "hecr")) |>
  sf::st_crs()

thalweg_points <- extract_thalweg_through_2D_domain(hc = h,
                      crs = crs_prj,
                      number_of_sections = 10,
                      verbose = TRUE)

#   do.call(dplyr::bind_rows,thalweg_points)|>
#  ggplot()+
#  geom_sf(aes(geometry = geom))
}
