% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-dim.R
\name{extract_ts1}
\alias{extract_ts1}
\title{Function retrieves a desired time series from a cross section part of a
HEC-RAS model}
\usage{
extract_ts1(.f, station_name, ts_type)
}
\arguments{
\item{.f}{an hdf5 file read in via hec_file or h5::h5file}

\item{station_name}{station for the cross section to query time series from}

\item{ts_type}{time series to query out (ex 'Water Surface', 'Depth', ...)}
}
\description{
Function retrieves a desired time series from a cross section part of a
HEC-RAS model
}
