% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-dim.R
\name{extract_ts2}
\alias{extract_ts2}
\title{Function extracts a time series from 2D portion of hec ras model.}
\usage{
extract_ts2(.f, xy, ts_type = "Water Surface", timestamp = NULL)
}
\arguments{
\item{.f}{an hdf file read in with hec_file or h5::h5file}

\item{xy}{a coordinate or set of coordinates either in a dataframe or matrix with columns x and y}

\item{ts_type}{the time series to extract, option defaults to 'Water Surface'}
}
\value{
dataframe for desired timeseries with relevant column attributes
}
\description{
Function extracts a time series from 2D portion of hec ras model.
}
