% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundary_condition_lines.R
\name{get_boundary_condition_lines}
\alias{get_boundary_condition_lines}
\title{Extract boundary condition lines from HEC-RAS hdf file}
\usage{
get_boundary_condition_lines(hc, crs, verbose = FALSE)
}
\arguments{
\item{hc}{an hdf file read in with the hec_file() function}

\item{crs}{character string (e.g., "EPSG:3035)}

\item{verbose}{boolean}
}
\value{
A sfc LINESTRING object
}
\description{
This function receives a hc object and crs and returns at least
  two boundary condition LINESTRINGs.
}
\examples{
path = system.file("extdata\\\\chippewa.hdf",package = "hecr")
h <- hecr::hec_file(path)

crs_prj <- sf::read_sf(system.file("extdata\\\\profile_lines.shp",
                                   package = "hecr")) |>
  sf::st_crs()

bc_lines <- get_boundary_condition_lines(hc = h,
                      crs = crs_prj,
                      verbose = TRUE)

#bc_lines |>
#  ggplot()+
#  geom_sf(aes(geometry = geom,
#  col = boundary_condition_pairs))
}
