% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_cell_centres.R
\name{get_comp_cell_centres}
\alias{get_comp_cell_centres}
\title{Extract computational cell cetres from 2d model domain}
\usage{
get_comp_cell_centres(hc, crs, verbose = FALSE)
}
\arguments{
\item{hc}{an hdf file read in with the hec_file() function}

\item{crs}{character string (e.g., "EPSG:3035)}

\item{verbose}{boolean}
}
\value{
A sfc POINT object containing all computational cell points to be
  joined to modelled parameters (e.g., WSE)
}
\description{
HEC-RAS exports computational cell centres to hdf, but also
  exports the perimeter cells, which are not appropriate to use for
  validation. This function identifies only the true cells fully within the
  model domain.
}
\details{
This function removes perimeter cells by using geos to spatially
  intersect cells with the model domain perimeter. This is potentially not
  the most efficient method, but it provides the required functionality.
}
\examples{
path = system.file("extdata\\\\chippewa.hdf",package = "hecr")
h <- hecr::hec_file(path)

crs_prj <- sf::read_sf(system.file("extdata\\\\profile_lines.shp",
                                   package = "hecr")) |>
  sf::st_crs()

comp_cell_centres <- get_comp_cell_centres(hc = h,
                      crs = crs_prj,
                      verbose = TRUE)

#comp_cell_centres |>
#  ggplot()+
#  geom_sf(aes(geometry = geom))
}
