% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{get_model_timestamps}
\alias{get_model_timestamps}
\title{Function retrieves the timestamps defined in the HEC-RAS model run}
\usage{
get_model_timestamps(.f)
}
\arguments{
\item{.f}{an hdf file read in with hec_file or h5::h5file}
}
\value{
a vector of datetimes
}
\description{
Function retrieves the timestamps defined in the HEC-RAS model run
}
