% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upstream_boundary_condition.R
\name{get_upstream_boundary_condition}
\alias{get_upstream_boundary_condition}
\title{Extract upstream hydrograph boundary condition}
\usage{
get_upstream_boundary_condition(hc, time_zone, verbose = FALSE)
}
\arguments{
\item{hc}{an hdf file read in with the hec_file() function}

\item{time_zone}{character string (e.g. "UTC)}

\item{verbose}{boolean}
}
\description{
HEC-RAS exports boundary condition hydrographs.
}
\details{
This function retrieves boundary condition time series related to
  upstream boundary condition lines.
}
\examples{
path = system.file("extdata\\\\Chippewa_2D.p05.hdf",package = "hecr")
h <- hecr::hec_file(path)

upstream_bc <- get_upstream_boundary_condition(hc = h,
                                               time_zone = "America/Chicago",
                      verbose = TRUE)

# upstream_bc |>
#  ggplot()+
#  geom_line(aes(x = datetime, y = discharge))
}
