% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-config.R
\name{hec_config}
\alias{hec_config}
\title{Function reads in and fills in a query based on a configuration file parameter}
\usage{
hec_config(config_file, config = "default")
}
\arguments{
\item{config_file}{path to config file used to fill query}

\item{config}{the config to use within the configuration file. Defaults to 'default'}
}
\value{
data frame
}
\description{
Function reads in and fills in a query based on a configuration file parameter
}

