% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R
\name{hec_datasets}
\alias{hec_datasets}
\title{List Available data}
\usage{
hec_datasets(hc, domain = NULL)
}
\arguments{
\item{hc}{a "hec" object}

\item{domain}{which domain to list data for, will default to both options are 'one' or 'two'}
}
\description{
list the data that is avaialble to query on. This values
depends on whether a query is set to 'one' or 'two' dimensional domains
}
