% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{hec_file}
\alias{hec_file}
\title{Function reads in an hdf5 file resulting from a HecRas model run. This function has been deprecated in favor of hec_file2}
\usage{
hec_file(f)
}
\arguments{
\item{f}{an hdf5 file resulting from a HecRas model run}
}
\description{
Function reads in an hdf5 file resulting from a HecRas model run. This function has been deprecated in favor of hec_file2
}
