% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{hec_file}
\alias{hec_file}
\title{Read HEC-Ras Output Data}
\usage{
hec_file(path, plan_numbers = NULL)
}
\arguments{
\item{path}{directory path to either a single hdf file or a directory of a collection of these}

\item{plan_numbers}{a vector of plan number associated with hdf files. For use when path is a directory.
Default action is to read all hdf files in path.}
}
\value{
a "hec_collection" object
}
\description{
Function will read in a single hdf file when the path parameter points to a single
file, or a collection of these when the path is a directory. User can further specify only a 
certain number of these within the directory to be read using the plan_numbers argument.
}
\examples{
\dontrun{
# read in a single hdf file
a <- hec_file("examples/ArdenwoodCreek.p50.hdf")

# read in a collection of hdf files in this directory
b <- hec_file("examples/") 

# read in hdf files in this directory matching 50 and 60 plan numbers
c <- hec_file("examples/", plan_numbers = c(50, 60))
}
}
