% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-file.R
\name{hec_file}
\alias{hec_file}
\title{Read hec file}
\usage{
hec_file(path, mode = "r")
}
\arguments{
\item{path}{string path to hdf file}
}
\value{
a hec object
}
\description{
read in a hdf file resulting from a hec-ras model run
}
