% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hec-utils.R, R/utils.R
\name{hec_info}
\alias{hec_info}
\alias{hec_info}
\title{Get Plan Information}
\usage{
hec_info(f)

hec_info(f)
}
\arguments{
\item{f}{a hec object}

\item{hc}{a set of hec files read in with hec_file()}
}
\value{
tibble with attributes as columns and unique plans as rows
}
\description{
Given a plan or a collection of these return a set of metadata attributes
defined for each plan.

get plan information for a given hec file or collection of these
}
